# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "CRM API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/owner/project/list": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "List Project",
                "description": "List Project",
                "operationId": "ProjectList",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "no of record you already get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "no of record you want to get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by name, mobile, email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column(eg. id, name, mobile, email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project/store": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Store a Project",
                "description": "Project",
                "operationId": "ProjectStore",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter Project Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Project Name"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Enter Project Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Project Code"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Enter Status(InPlanning, Running, Stopped, Completed)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Active"
                    },
                    {
                        "name": "total_hours",
                        "in": "query",
                        "description": "Enter Total Hours",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "100"
                    },
                    {
                        "name": "per_day_hours",
                        "in": "query",
                        "description": "Enter Per Day Hours",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "8"
                    },
                    {
                        "name": "assigned_user",
                        "in": "query",
                        "description": "Enter Assigned User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1,2,3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project/update": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Update a Project",
                "description": "Project",
                "operationId": "ProjectUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter Project Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Project Name"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Enter Project Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Project Code"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Enter Status(InPlanning, Running, Stopped, Completed)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Active"
                    },
                    {
                        "name": "total_hours",
                        "in": "query",
                        "description": "Enter Total Hours",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "100"
                    },
                    {
                        "name": "per_day_hours",
                        "in": "query",
                        "description": "Enter Per Day Hours",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "8"
                    },
                    {
                        "name": "assigned_user",
                        "in": "query",
                        "description": "Enter Assigned User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1,2,3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project/view": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "View a Project",
                "description": "Project",
                "operationId": "ProjectView",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project/change-status": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Change Project Status",
                "description": "Project",
                "operationId": "ProjectChangeStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Enter Status(InPlanning, Running, Stopped, Completed)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "InPlanning",
                                "Running",
                                "Stopped",
                                "Completed"
                            ]
                        },
                        "example": "Active"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project/destroy": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Destroy a Project",
                "description": "Project",
                "operationId": "ProjectDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Project ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/project/drop-down": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Get Project Drop Down",
                "description": "Project",
                "operationId": "ProjectDropDown",
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}